+++ logkitweb/public/manifest.json
