+++ logkitweb/src/createContainer.js
@@ -1,9 +1,10 @@
 import React, {Component} from 'react';
 import {notification, message, Button, Steps} from 'antd';
-import Source from  './containers/sourceConfig'
-import Parser from  './containers/parserConfig'
-import Sender from './containers/senderConfig'
-import RenderConfig from './containers/renderConfig'
+import Source from  './components/sourceConfig'
+import Parser from  './components/parserConfig'
+import Sender from './components/senderConfig'
+import RenderConfig from './components/renderConfig'
+
 
 const Step = Steps.Step;
 const steps = [{
@@ -29,8 +30,9 @@ class Create extends Component {
   }
 
   componentDidMount() {
-
+    this.init()
   }
+
 
   componentWillUnmount() {
 
@@ -38,6 +40,15 @@ class Create extends Component {
 
   componentDidUpdate(prevProps) {
 
+  }
+
+  init = () => {
+    let isCopy =  this.props.location.query.copyConfig
+    if (isCopy == 'true') {
+      this.setState({
+        current: 3
+      })
+    }
   }
 
   next() {
@@ -81,10 +92,13 @@ class Create extends Component {
     this.setState({current});
   }
 
+  turnToIndex() {
+    this.props.router.push({pathname: `/index`})
+  }
+
 
   render() {
     const {current} = this.state;
-    console.log(this.state.current)
     return (
         <div className="logkit-create-container">
           <div className="header">七牛Logkit配置文件助手</div>
@@ -119,7 +133,7 @@ class Create extends Component {
             {
               this.state.current === steps.length - 1
               &&
-              <Button type="primary" onClick={() => message.success('Processing complete!')}>Done</Button>
+              <Button type="primary" onClick={() => this.turnToIndex()}>回到首页</Button>
             }
             {
               this.state.current > 0
