+++ logkitweb/src/services/logkit.js
@@ -1,7 +1,9 @@
 /**
- * Created by zhonghuiping on 2017/7/7.
+ * Created by zhonghuiping on 2017/8/25.
  */
 import request  from '../utils/request';
+
+window.logkitUrl = ''
 
 /* window.logkitUrl = 'http://localhost:12581' */
 
@@ -30,7 +32,6 @@ export async function getSourceParseOptions(params) {
 }
 
 export async function getSourceParsesamplelogs(params) {
-  window.logkitUrl = 'http://localhost:12581'
   return request(window.logkitUrl + '/logkit/parser/samplelogs', {
     method: 'get',
   });
@@ -49,6 +50,12 @@ export async function getSenderOptions(params) {
 }
 
 export async function getRunnerConfigs(params) {
+  return request(window.logkitUrl + '/logkit/configs', {
+    method: 'get',
+  });
+}
+
+export async function getRunnerStatus(params) {
   return request(window.logkitUrl + '/logkit/status', {
     method: 'get',
   });
