+++ logkitweb/src/utils/request.js
@@ -25,8 +25,6 @@ export default function request(url, options) {
     .then(checkStatus)
     .then(parseJSON)
     .then((data) => {
-      console.log(data)
-      data.success = true;
       if(data.message != undefined) {
         notification.error({message: "失败", description: data.message, duration: 20})
         data.success = false;
