--- public-old/v2/logkitweb/README.md
@@ -1,7 +1,0 @@
-#author zhonghuiping
-
-# init project
-yarn install
-
-# start
-yarn statr
